evaluate_cbr/pyproject.toml
[tool.poetry]
name = "evaluate-cbr"
version = "0.1.0"
description = ""
authors = ["Kilian Bartz <kilianbartz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.1.1"
torchvision = "^0.16.1"
datasets = "^2.15.0"
transformers = "^4.35.2"
accelerate = "^0.24.1"
huggingface-hub = "^0.19.4"
matplotlib = "^3.8.2"
faiss-cpu = "^1.7.4"
arg-services = "^1.6.0"
arguebuf = "^2.5.0"
pybars3 = "^0.9.7"
jupyterlab = "^4.0.9"
wandb = "^0.16.1"
jupyter = "^1.0.0"
sentence-transformers = "^2.2.2"
pytorch-metric-learning = "^2.3.0"
python-dotenv = "^1.0.0"
openai = "^1.6.1"
langchain = "^0.0.353"
byol-pytorch = "^0.8.0"
torchmetrics = "^1.2.1"
setuptools = "^69.0.3"
lightning = "^2.1.3"
pytorch-lightning = "^2.1.3"
urllib3 = "^2.1.0"
seaborn = "^0.13.1"
ranx = "^0.3.19"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
